타이머 UI 로직 수정 및 00초마다 빨간색으로 수정

+++ DeathRun/Assets/Scripts/System/LevelData.cs
@@ -12,7 +12,7 @@ public class LevelData : MonoBehaviour
     private bool isGameStart = false;
 
     [SerializeField] private int startPlayerCount;
-    [SerializeField] private float LimitTime;
+    [SerializeField] private int LimitTime;
 
     public GameObject[] savePoints;
     public Image savePointImage;
